Pipe animation polish + Implementation of a scoring system
I polished the pipe animation so that there is an infinite number of pipes that spawn in random heights as well as that game object being destroyed to preserve more memory. Finally I implemented a scoring system that will be integrated into the game, once the collision system is put into place.